refactor(i18n): move transifex api to typescript module
