+++ package.json
@@ -65,7 +65,6 @@
     "packageExports": true
   },
   "dependencies": {
-    "@transifex/api": "7.1.3",
     "axios": "1.7.7",
     "axios-retry": "4.5.0",
     "chalk": "5.3.0",
