+++ packages/i18n/package.json
@@ -4,12 +4,16 @@
   "description": "localization interfaces",
   "type": "module",
   "main": "build/index.js",
+  "types": "build/index.d.ts",
   "scripts": {
-    "build": "tsc --project tsconfig.json --outDir build",
+    "build": "tsc --project tsconfig.json",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "AGPL-3.0-or-later",
+  "dependencies": {
+    "@transifex/api": "7.1.3"
+  },
   "devDependencies": {
     "chalk": "5.3.0"
   }
