+++ packages/i18n/src/index.ts
@@ -1,1 +1,2 @@
-export { default } from './locales.js'
+export { default, type LocaleDefinition, type LocaleLevel } from './locales.js'
+export { getFromTransifex } from './transifex.js'
