+++ packages/types/index.d.ts
@@ -143,16 +143,6 @@ export type UnitsSetting =
   | typeof SETTINGS_UNITS_METRIC
   | typeof SETTINGS_UNITS_IMPERIAL
 
-export type LocaleLevel = 1 | 2 | 3 | 4
-
-export interface LocaleDefinition {
-  label: string
-  name: string
-  value: string
-  key: string
-  level: LocaleLevel
-}
-
 export type BuildingPosition = 'left' | 'right'
 
 // Subset of / derived from SegmentDefinition
